generate per-tenant urlconf in sub-folder middleware
This fixes django-tenants/django-tenants#820 until a better method is
found.

The root cause is that information stored on connections is not
available on all contexts during the processing of a request. In WSGI
mode a single thread is involved so this is not visible. In ASGI mode
however some code execute in the context of the async event loop. This
is because the connection handler stores information in a
thread_critical variable as seen here https://github.com/django/django/blob/main/django/utils/connection.py#L41.

A better method would be to use asgi.local with our own sharing policy
instead instead of attaching to the connexion object.

Until that is settled (@lorinkoz has a prototype in
https://github.com/lorinkoz/django-pgschemas/pull/120) this patch
modifies urlconf management in the subpath middleware.

IO/connection operations which were previously done during resolve() are
removed, which prevents the error from popping up in ASGI mode

The downside is that we're modifying that we're generating a new urlconf
object for each tenant. Also, this fixes a particular subpath issue, but
does not address the more general issue of using connection to store a
state accross each request processing.